pyproject.toml
[project]
name = "olist-ml"
version = "0.1.0"
description = ""
authors = [
    {name = "olverarcedev",email = "olverarcedev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "numpy (>=2.4.0,<3.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "jupyterlab (>=4.5.1,<5.0.0)",
    "ruff (>=0.14.10,<0.15.0)"
]

[tool.poetry]
packages = [{include = "olist_ml", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "hydra-colorlog (>=1.2.0,<2.0.0)",
    "dvc[gs] (>=3.65.0,<4.0.0)",
    "mkdocs (>=1.6.1,<2.0.0)"
]


[tool.poetry.scripts]
run-pipeline = "olist_ml.main:main"
prepare-data = "olist_ml.data:main"